+++ hutool-extra/src/test/java/cn/hutool/extra/pinyin/PinyinUtilTest.java
@@ -1,5 +1,6 @@
 package cn.hutool.extra.pinyin;
 
+import cn.hutool.extra.pinyin.engine.bopomofo4j.Bopomofo4jEngine;
 import cn.hutool.extra.pinyin.engine.pinyin4j.Pinyin4jEngine;
 import org.junit.Assert;
 import org.junit.Test;
@@ -20,6 +21,13 @@ public class PinyinUtilTest {
 	}
 
 	@Test
+	public void getPinyinByBopomofo4jTest() {
+		final Bopomofo4jEngine engine = new Bopomofo4jEngine();
+		final String pinyin = engine.getPinyin("你好h", " ");
+		Assert.assertEquals("ni haoh", pinyin);
+	}
+
+	@Test
 	public void getPinyinUpperCaseTest(){
 		final String pinyin = PinyinUtil.getPinyin("你好怡", " ");
 		Assert.assertEquals("ni hao yi", pinyin);
@@ -34,6 +42,13 @@ public class PinyinUtilTest {
 	@Test
 	public void getFirstLetterByPinyin4jTest(){
 		final Pinyin4jEngine engine = new Pinyin4jEngine();
+		final String result = engine.getFirstLetter("林海", "");
+		Assert.assertEquals("lh", result);
+	}
+
+	@Test
+	public void getFirstLetterByBopomofo4jTest(){
+		final Bopomofo4jEngine engine = new Bopomofo4jEngine();
 		final String result = engine.getFirstLetter("林海", "");
 		Assert.assertEquals("lh", result);
 	}